Update build.py
Hello,
I found that on my system, the dirname of __file__ returns nothing, therefore the build program is trying to install the firmware dir on the root. Which could happen to those that entered the password for the previous sudo installations.
This correction is safe and the software keep compatibility to other system that were working.
Thanks for your precious work.